requirements.txt
pydiscourse>=1.3,<1.4
PyYAML>=6.0,<6.1
requests>=2.28,<2.29
GitPython>=3.1,<3.2
PyGithub>=1.57,<1.58
